requirements.txt
langchain
langchain_core
langchain_google_genai
streamlit
python-dotenv
langserve[all]
uvicorn
fastapi